Fixed subtle error that influenced the Variability calculation

+++ paint-generate-squares/src/main/java/generatesquares/calc/CalculateVariability.java
@@ -87,8 +87,8 @@ public class CalculateVariability {
                                     int nrOfSquaresInRow,
                                     int granularity) {
         // Calculate the top-left corner (x0, y0) of the square
-        double x0 = ((double)  squareSeqNr % nrOfSquaresInRow) * width;
-        double y0 = ((double) squareSeqNr / nrOfSquaresInRow) * height;
+        double x0 = (squareSeqNr % nrOfSquaresInRow) * width;
+        double y0 = (squareSeqNr / nrOfSquaresInRow) * height;
 
         // Calculate the grid indices (xi, yi) for the track
         int xi = (int) (((x1 - x0) / width) * granularity);
